+++ src/aux-server/aux-backend/prisma/PrismaAuthStore.ts
@@ -914,78 +914,94 @@ export class PrismaAuthStore implements AuthStore {
         const periodStart = convertToDate(request.currentPeriodStartMs);
         const periodEnd = convertToDate(request.currentPeriodEndMs);
         if (request.userId) {
+            const updateData: Prisma.UserUpdateArgs<any>['data'] = {
+                subscriptionId: request.subscriptionId,
+                subscriptionStatus: request.subscriptionStatus,
+                subscriptionPeriodStart: periodStart,
+                subscriptionPeriodEnd: periodEnd,
+            };
+
+            if (request.stripeCustomerId) {
+                updateData.stripeCustomerId = request.stripeCustomerId;
+            }
+
+            if (request.stripeSubscriptionId) {
+                updateData.subscriptionInfo = {
+                    upsert: {
+                        create: {
+                            id: uuid(),
+                            userId: request.userId,
+                            subscriptionId: request.subscriptionId,
+                            subscriptionStatus: request.subscriptionStatus,
+                            stripeCustomerId: request.stripeCustomerId,
+                            stripeSubscriptionId: request.stripeSubscriptionId,
+                            currentPeriodStart: periodStart,
+                            currentPeriodEnd: periodEnd,
+                        },
+                        update: {
+                            subscriptionId: request.subscriptionId,
+                            subscriptionStatus: request.subscriptionStatus,
+                            stripeCustomerId: request.stripeCustomerId,
+                            stripeSubscriptionId: request.stripeSubscriptionId,
+                            currentPeriodStart: periodStart,
+                            currentPeriodEnd: periodEnd,
+                        },
+                    },
+                };
+            } else {
+                updateData.subscriptionInfoId = null;
+            }
+
             await this._client.user.update({
                 where: {
                     id: request.userId,
                 },
-                data: {
-                    subscriptionId: request.subscriptionId,
-                    subscriptionStatus: request.subscriptionStatus,
-                    stripeCustomerId: request.stripeCustomerId,
-                    subscriptionPeriodStart: periodStart,
-                    subscriptionPeriodEnd: periodEnd,
-                    subscriptionInfo: {
-                        upsert: {
-                            create: {
-                                id: uuid(),
-                                userId: request.userId,
-                                subscriptionId: request.subscriptionId,
-                                subscriptionStatus: request.subscriptionStatus,
-                                stripeCustomerId: request.stripeCustomerId,
-                                stripeSubscriptionId:
-                                    request.stripeSubscriptionId,
-                                currentPeriodStart: periodStart,
-                                currentPeriodEnd: periodEnd,
-                            },
-                            update: {
-                                subscriptionId: request.subscriptionId,
-                                subscriptionStatus: request.subscriptionStatus,
-                                stripeCustomerId: request.stripeCustomerId,
-                                stripeSubscriptionId:
-                                    request.stripeSubscriptionId,
-                                currentPeriodStart: periodStart,
-                                currentPeriodEnd: periodEnd,
-                            },
-                        },
-                    },
-                },
+                data: updateData,
             });
         } else if (request.studioId) {
+            const updateData: Prisma.StudioUpdateArgs<any>['data'] = {
+                subscriptionId: request.subscriptionId,
+                subscriptionStatus: request.subscriptionStatus,
+                subscriptionPeriodStart: periodStart,
+                subscriptionPeriodEnd: periodEnd,
+            };
+
+            if (request.stripeCustomerId) {
+                updateData.stripeCustomerId = request.stripeCustomerId;
+            }
+
+            if (request.stripeSubscriptionId) {
+                updateData.subscriptionInfo = {
+                    upsert: {
+                        create: {
+                            id: uuid(),
+                            studioId: request.studioId,
+                            subscriptionId: request.subscriptionId,
+                            subscriptionStatus: request.subscriptionStatus,
+                            stripeCustomerId: request.stripeCustomerId,
+                            stripeSubscriptionId: request.stripeSubscriptionId,
+                            currentPeriodStart: periodStart,
+                            currentPeriodEnd: periodEnd,
+                        },
+                        update: {
+                            subscriptionId: request.subscriptionId,
+                            subscriptionStatus: request.subscriptionStatus,
+                            stripeCustomerId: request.stripeCustomerId,
+                            stripeSubscriptionId: request.stripeSubscriptionId,
+                            currentPeriodStart: periodStart,
+                            currentPeriodEnd: periodEnd,
+                        },
+                    },
+                };
+            } else {
+                updateData.subscriptionInfoId = null;
+            }
+
             await this._client.studio.update({
                 where: {
                     id: request.studioId,
                 },
-                data: {
-                    subscriptionId: request.subscriptionId,
-                    subscriptionStatus: request.subscriptionStatus,
-                    stripeCustomerId: request.stripeCustomerId,
-                    subscriptionPeriodStart: periodStart,
-                    subscriptionPeriodEnd: periodEnd,
-                    subscriptionInfo: {
-                        upsert: {
-                            create: {
-                                id: uuid(),
-                                studioId: request.studioId,
-                                subscriptionId: request.subscriptionId,
-                                subscriptionStatus: request.subscriptionStatus,
-                                stripeCustomerId: request.stripeCustomerId,
-                                stripeSubscriptionId:
-                                    request.stripeSubscriptionId,
-                                currentPeriodStart: periodStart,
-                                currentPeriodEnd: periodEnd,
-                            },
-                            update: {
-                                subscriptionId: request.subscriptionId,
-                                subscriptionStatus: request.subscriptionStatus,
-                                stripeCustomerId: request.stripeCustomerId,
-                                stripeSubscriptionId:
-                                    request.stripeSubscriptionId,
-                                currentPeriodStart: periodStart,
-                                currentPeriodEnd: periodEnd,
-                            },
-                        },
-                    },
-                },
+                data: updateData,
             });
         }
     }