+++ src/aux-server/aux-web/shared/scene/decorators/DimensionPositionDecorator.ts
@@ -1,31 +1,20 @@
-import { AuxBot3DDecorator, AuxBot3DDecoratorBase } from '../AuxBot3DDecorator';
+import { AuxBot3DDecoratorBase } from '../AuxBot3DDecorator';
 import { AuxBot3D } from '../AuxBot3D';
 import {
     calculateNumericalTagValue,
     BotCalculationContext,
     Bot,
-    calculateGridScale,
-    objectsAtDimensionGridPosition,
     getBotPosition,
     getBotRotation,
-    getDimensionScale,
-    getDimensionGridHeight,
     cacheFunction,
-    calculateBooleanTagValue,
     getBotOrientationMode,
-    getBotAnchorPoint,
-    BotAnchorPoint,
     BotOrientationMode,
     getBotIndex,
-    getBotScale,
     getAnchorPointOffset,
     LocalActions,
-    Easing,
-    EaseMode,
     hasValue,
     enqueueAsyncResult,
     BotAction,
-    LocalTweenAction,
     LocalRotationTweenAction,
     LocalPositionTweenAction,
     enqueueAsyncError,
@@ -34,32 +23,20 @@ import {
 } from '@casual-simulation/aux-common';
 import {
     Rotation,
-    AUX_ROTATION_TO_THREEJS,
     Vector3 as CasualVector3,
-    LookRotation,
 } from '@casual-simulation/aux-common/math';
 import {
     Vector3,
     Quaternion,
     Euler,
-    Vector2,
     Object3D,
     MathUtils as ThreeMath,
     Matrix4,
 } from '@casual-simulation/three';
 import { calculateGridTileLocalCenter } from '../grid/Grid';
-import { realPosToGridPos, Axial, posToKey } from '../hex';
-import { BuilderGroup3D } from '../BuilderGroup3D';
-import {
-    calculateScale,
-    cameraUpwardRay,
-    objectForwardRay,
-    WORLD_UP,
-} from '../SceneUtils';
+import { calculateScale } from '../SceneUtils';
 import { Game } from '../Game';
-import TWEEN, { Tween } from '@tweenjs/tween.js';
-import { MapSimulation3D } from '../../../aux-player/scene/MapSimulation3D';
-import { CoordinateSystem } from '../CoordinateSystem';
+import TWEEN from '@tweenjs/tween.js';
 import { DimensionGroup3D } from '../DimensionGroup3D';
 
 const tempVector3 = new Vector3();
@@ -163,32 +140,6 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
                 coordinateTransform
             );
 
-            if (
-                this._positionUpdated(currentGridPos, currentSortOrder) ||
-                this._heightUpdated(currentHeight)
-            ) {
-                let ids = [] as string[];
-                if (this._lastPos) {
-                    const objectsAtLastPosition =
-                        objectsAtDimensionGridPosition(
-                            calc,
-                            this.bot3D.dimension,
-                            this._lastPos
-                        );
-                    ids.push(...objectsAtLastPosition.map((b) => b.id));
-                }
-                if (currentGridPos) {
-                    const objectsAtCurrentPosition =
-                        objectsAtDimensionGridPosition(
-                            calc,
-                            this.bot3D.dimension,
-                            currentGridPos
-                        );
-                    ids.push(...objectsAtCurrentPosition.map((b) => b.id));
-                }
-
-                this.bot3D.dimensionGroup.simulation3D.ensureUpdate(ids);
-            }
             this._lastPos = currentGridPos;
             this._lastSortOrder = currentSortOrder;
             this._lastHeight = currentHeight;
@@ -231,18 +182,7 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
         }
     }
 
-    botRemoved(calc: BotCalculationContext): void {
-        if (this._lastPos) {
-            const objectsAtPosition = objectsAtDimensionGridPosition(
-                calc,
-                this.bot3D.dimension,
-                this._lastPos
-            );
-            this.bot3D.dimensionGroup.simulation3D.ensureUpdate(
-                objectsAtPosition.map((f) => f.id)
-            );
-        }
-    }
+    botRemoved(calc: BotCalculationContext): void {}
 
     private _heightUpdated(currentHeight: number): boolean {
         return Math.abs(this._lastHeight - currentHeight) > 0.01;