+++ src/aux-vm-browser/managers/SystemPortalCoordinator.ts
@@ -46,7 +46,6 @@ import {
     SystemPortalPane,
     getSystemPortalPane,
     getOpenSystemPortalPane,
-    updatedBot,
     isModule,
     parseModule,
 } from '@casual-simulation/aux-common';